Extract static helper functions
