+++ Backend/Helper/FileOperations.cs
@@ -1,8 +1,18 @@
-﻿using System.IO;
+﻿using System;
+using System.IO;
 using System.IO.Compression;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
 
 namespace BackendFramework.Helper
 {
+    [Serializable]
+    public class InvalidFileException : Exception
+    {
+        public InvalidFileException(string message) : base(message) { }
+
+    }
+
     /// <summary>
     /// A collection of general purpose file system operations.
     /// </summary>
@@ -43,6 +53,31 @@ namespace BackendFramework.Helper
                 File.Delete(zipFilePath);
             }
             return extractionDir;
+        }
+
+        /// <summary> Extract a zip file to a unique temporary directory. </summary>
+        /// <param name="file"> An IFormFile zip file. </param>
+        /// <returns> The path to the extracted contents. </returns>
+        public static async Task<string> ExtractZipFile(IFormFile? file)
+        {
+            if (file is null)
+            {
+                throw new InvalidFileException("Null file");
+            }
+            if (file.Length == 0)
+            {
+                throw new InvalidFileException("Empty file");
+            }
+
+            // Copy zip file data to a new temporary file
+            var filePath = Path.GetTempFileName();
+            await using (var fs = new FileStream(filePath, FileMode.OpenOrCreate))
+            {
+                await file.CopyToAsync(fs);
+            }
+
+            // Extract the zip to new created directory.
+            return ExtractZipFile(filePath, null, true);
         }
 
         /// <summary>
